refactor(api): move dict_from_unifi_list function to convert module
